The .rk file is now using the _on_say method correctly
A working example of how to read a .rk file into the template now functions.

Once we have a more complex .rk file we can develop how each use case of an .rk script executes.
Also removed the changes in the Game.gd, this was the incorrect file.

To access the code that is running via the .rk file:
Click "Play"
Click "Dynamic"

Important note for exporting:
https://rakugoteam.github.io/rakugo-docs/2.2/export/